# InternPool: introduce TrackedInst to prepare for incremental compilation