… src/AstGen.zig
… src/Module.zig
… src/Sema.zig
… src/Zir.zig
… src/main.zig
… src/print_zir.zig