See commit messages for details. I will provide performance data points shortly.