@mlugg 2024-01-23
Oh, just to document a change that I snuck into the first commit, our ZIR mapping algorithm is now more effective in that we no longer ignore unnamed decls and tests. These were previously just never mapped, which in practice would have resulted in re-analysis on every update!