@mlugg 2024-01-23
## Building behavior tests with x86_64 backend:
```
Benchmark 1 (31 runs): /home/mlugg/zig/master/build/stage3/bin/zig test -I/home/mlugg/zig/master/test /home/mlugg/zig/master/test/behavior.zig -fno-llvm -fno-lld --test-no-exec --cache-dir c1 --global-cache-dir c2
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          1.62s  ±  142ms    1.43s  … 1.87s           0 ( 0%)        0%
  peak_rss            121MB ±  200KB     120MB …  121MB          0 ( 0%)        0%
  cpu_cycles         4.98G  ± 4.06M     4.98G  … 4.99G           0 ( 0%)        0%
  instructions       6.51G  ± 25.3K     6.51G  … 6.51G           2 ( 6%)        0%
  cache_references    398M  ± 1.66M      395M  …  404M           1 ( 3%)        0%
  cache_misses       91.9M  ±  358K     91.2M  … 92.7M           3 (10%)        0%
  branch_misses      41.9M  ± 24.6K     41.8M  … 41.9M           0 ( 0%)        0%
Benchmark 2 (31 runs): /home/mlugg/zig/incremental-the-second/stage4-release/bin/zig test -I/home/mlugg/zig/master/test /home/mlugg/zig/master/test/behavior.zig -fno-llvm -fno-lld --test-no-exec --cache-dir c3 --global-cache-dir c4
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          1.63s  ±  160ms    1.45s  … 1.99s           0 ( 0%)          +  0.6% ±  4.8%
  peak_rss            123MB ±  168KB     123MB …  123MB          0 ( 0%)        💩+  2.1% ±  0.1%
  cpu_cycles         5.02G  ± 3.02M     5.01G  … 5.02G           1 ( 3%)          +  0.7% ±  0.0%
  instructions       6.51G  ± 24.7K     6.51G  … 6.51G           0 ( 0%)          -  0.1% ±  0.0%
  cache_references    406M  ± 1.52M      403M  …  411M           1 ( 3%)        💩+  2.0% ±  0.2%
  cache_misses       93.9M  ±  473K     92.8M  … 95.1M           2 ( 6%)        💩+  2.2% ±  0.2%
  branch_misses      42.6M  ± 19.5K     42.6M  … 42.7M           2 ( 6%)        💩+  1.8% ±  0.0%
```
Peak RSS increased slightly, which is to be expected, and led to a few more cache references, but the resulting performance delta is insignificant.

---
## `AstGen` on `Sema.zig`:
```
Benchmark 1 (461 runs): /home/mlugg/zig/master/build/stage3/bin/zig ast-check /home/mlugg/zig/master/src/Sema.zig
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           108ms ± 6.28ms     104ms …  130ms         48 (10%)        0%
  peak_rss           64.8MB ± 28.5KB    64.7MB … 64.8MB         45 (10%)        0%
  cpu_cycles          342M  ± 3.34M      336M  …  357M           8 ( 2%)        0%
  instructions        550M  ± 77.7       550M  …  550M         107 (23%)        0%
  cache_references   8.51M  ±  152K     8.13M  … 9.42M          12 ( 3%)        0%
  cache_misses        514K  ± 19.8K      476K  …  619K          15 ( 3%)        0%
  branch_misses      4.01M  ± 3.29K     4.00M  … 4.02M           2 ( 0%)        0%
Benchmark 2 (472 runs): /home/mlugg/zig/incremental-the-second/stage4-release/bin/zig ast-check /home/mlugg/zig/master/src/Sema.zig
  measurement          mean ± σ            min … max           outliers         delta
  wall_time           106ms ± 4.98ms     103ms …  128ms         36 ( 8%)        ⚡-  2.3% ±  0.7%
  peak_rss           64.7MB ± 14.7KB    64.4MB … 64.7MB          5 ( 1%)          -  0.2% ±  0.0%
  cpu_cycles          335M  ± 1.53M      333M  …  341M           8 ( 2%)        ⚡-  2.1% ±  0.1%
  instructions        550M  ± 78.1       550M  …  550M         112 (24%)          -  0.0% ±  0.0%
  cache_references   8.98M  ±  144K     8.63M  … 9.71M          10 ( 2%)        💩+  5.6% ±  0.2%
  cache_misses        588K  ± 28.7K      523K  …  722K           7 ( 1%)        💩+ 14.3% ±  0.6%
  branch_misses      3.95M  ± 7.70K     3.93M  … 3.97M           0 ( 0%)        ⚡-  1.5% ±  0.0%
```
Interestingly, there's a big jump in cache misses, but an overall performance improvement! This probably comes from the fact that we're not doing as much bit fiddling when constructing declarations anymore.

---
## ZIR size of `Sema.zig`
| Branch | Total ZIR bytes |
|------------|-----------------------|
| `master` | `4.851020812988281MiB` |
| PR | `4.848020553588867MiB` |

That's an insignificant change in ZIR bytes; a reduction of less than 0.1%.