@andrewrk 2024-01-23
Yes please let me review this. My immediate reaction is that I think you have once again tried to add something inappropriate to InternPool, but I will keep an open mind and fully understand these changes before complaining about it.