@mlugg 2024-01-23
Okay - let me know if there's anything you would like me to clarify. FWIW, the sole purpose of this change is to *reduce* the amount of unstable (as in, must be modified across incremental updates) data in the InternPool, so I believe it is appropriate. (Garbage will be relatively low; garbage values will appear only when ZIR mapping fails to correlate an instruction, such as when a declaration is deleted, and in future I suspect we may be able to reuse these dead indices without a full GC sweep.)