@andrewrk 2024-01-23
One more data point that would be interesting would be the size of a serialized InternPool before and after this change for e.g. building the compiler. You can obtain this with the `--debug-incremental` flag. Note that this will require additional logic in `Compilation.saveState` since there is a new array added to InternPool.