@mlugg 2024-01-23
Sure, I'm happy to introduce that logic and check the size of this state. Would you like that to be ready before you review?