@mlugg 2024-01-23
Serialized InternPool state is 2.83 MiB before, 2.95 MiB after; 4% increase. I will push these changes shortly.