@mlugg 2024-01-23
Serialized InternPool state for building the compiler itself is 36.11 MiB before, 36.41 MiB after; 0.8% increase.