@andrewrk 2024-01-24
Can you provide a demonstration of running `zig changelist` on some examples to show how these changes produce better or equivalent results?